+++ tests/test_model.py
@@ -62,8 +62,8 @@ def test_test_step(sample_config):
     )
     sample_batch = sample_input.unsqueeze(0).repeat(2, 1)
     model.test_step((sample_batch, sample_batch), 0)
-    assert len(model.test_generated[0.1]) == 2
-    assert len(model.test_real[0.1]) == 2
+    assert len(model.test_generated[0.3]) == 2
+    assert len(model.test_real[0.3]) == 2
     model.on_test_epoch_end()